+++ src/main/scala/perflex/statkind/Time.scala
@@ -1,5 +1,5 @@
 package perflex.statkind
 
 trait Time {
-  def time: Long
+  def time: Float
 }
